+++ R/function_visualizations.R
@@ -41,7 +41,7 @@
 #' @export
 #'
 #' @examples
-#' library(msdata)
+#' library(MsDataHub)
 #' library(MsExperiment)
 #' library(S4Vectors)
 #' msexp <- MsExperiment()
@@ -52,12 +52,13 @@
 #'
 #' ## define file names containing spectra data for the samples and
 #' ## add them, along with other arbitrary files to the experiment
-#' fls <- dir(system.file("sciex", package = "msdata"), full.names = TRUE)
+#' sciex_file <- c(MsDataHub::X20171016_POOL_POS_1_105.134.mzML(),
+#'     MsDataHub::X20171016_POOL_POS_3_105.134.mzML())
 #'
 #'
 #' library(Spectra)
 #' ## import the data and add it to the msexp object
-#' spectra(msexp) <- Spectra(fls, backend = MsBackendMzR())
+#' spectra(msexp) <- Spectra(sciex_file)
 #'
 #'
 #' ## define the quality metrics to be calculated
@@ -137,7 +138,7 @@ plotMetric <- function(qc, metric = "areaUnderTic", plotly = TRUE) {
 #' @export
 #'
 #' @examples
-#' library(msdata)
+#' library(MsDataHub)
 #' library(MsExperiment)
 #' library(S4Vectors)
 #' msexp <- MsExperiment()
@@ -148,10 +149,14 @@ plotMetric <- function(qc, metric = "areaUnderTic", plotly = TRUE) {
 #'
 #' ## define file names containing spectra data for the samples and
 #' ## add them, along with other arbitrary files to the experiment
-#' fls <- dir(system.file("sciex", package = "msdata"), full.names = TRUE)
+#' ## define file names containing spectra data for the samples
+#' sciex_file <- c(MsDataHub::X20171016_POOL_POS_1_105.134.mzML(),
+#'     MsDataHub::X20171016_POOL_POS_3_105.134.mzML())
+#'
 #' experimentFiles(msexp) <- MsExperimentFiles(
-#'     mzML_files = fls,
+#'     mzML_files = sciex_file,
 #'     annotations = "internal_standards.txt")
+#'
 #' ## link samples to data files: first sample to first file in "mzML_files",
 #' ## second sample to second file in "mzML_files"
 #' msexp <- linkSampleData(msexp, with = "experimentFiles.mzML_files",
@@ -161,7 +166,7 @@ plotMetric <- function(qc, metric = "areaUnderTic", plotly = TRUE) {
 #'
 #' library(Spectra)
 #' ## import the data and add it to the mse object
-#' spectra(msexp) <- Spectra(fls, backend = MsBackendMzR())
+#' spectra(msexp) <- Spectra(sciex_file)
 #'
 #'
 #' ## define the quality metrics to be calculated
@@ -251,7 +256,7 @@ plotMetricTibble <- function(qc, metric) {
 #'
 #'
 #' @examples
-#' library(msdata)
+#' library(MsDataHub)
 #' library(MsExperiment)
 #' library(S4Vectors)
 #' msexp <- MsExperiment()
@@ -262,9 +267,12 @@ plotMetricTibble <- function(qc, metric) {
 #'
 #' ## define file names containing spectra data for the samples and
 #' ## add them, along with other arbitrary files to the experiment
-#' fls <- dir(system.file("sciex", package = "msdata"), full.names = TRUE)
+#' ## define file names containing spectra data for the samples
+#' sciex_file <- c(MsDataHub::X20171016_POOL_POS_1_105.134.mzML(),
+#'     MsDataHub::X20171016_POOL_POS_3_105.134.mzML())
+#'
 #' experimentFiles(msexp) <- MsExperimentFiles(
-#'     mzML_files = fls,
+#'     mzML_files = sciex_file,
 #'     annotations = "internal_standards.txt")
 #' ## link samples to data files: first sample to first file in "mzML_files",
 #' ## second sample to second file in "mzML_files"
@@ -276,7 +284,7 @@ plotMetricTibble <- function(qc, metric) {
 #'
 #' library(Spectra)
 #' ## import the data and add it to the mse object
-#' spectra(msexp) <- Spectra(fls, backend = MsBackendMzR())
+#' spectra(msexp) <- Spectra(sciex_file)
 #'
 #'
 #' ## define the quality metrics to be calculated