+++ tests/testthat/test_function_Spectra_qualityMetrics.R
@@ -1,15 +1,17 @@
 ## create toy example data set (Spectra)
-library(msdata)
+library(MsDataHub)
 library(Spectra)
 library(Chromatograms)
-fls <- dir(system.file("sciex", package = "msdata"), full.names = TRUE)
-spectra <- Spectra(fls, backend = MsBackendMzR())
-chr <- Chromatograms(spectra, "sum")
-
-## create toy example data set (MsExperiment)
-library(msdata)
 library(MsExperiment)
 library(S4Vectors)
+## define file names containing spectra data for the samples
+sciex_file <- c(MsDataHub::X20171016_POOL_POS_1_105.134.mzML(),
+    MsDataHub::X20171016_POOL_POS_3_105.134.mzML())
+
+## import the data and assign it to the spectra object
+spectra <- Spectra(files = sciex_file)
+chr <- Chromatograms(spectra, "sum")
+## create toy example data set (MsExperiment)
 msexp <- MsExperiment()
 sd <- DataFrame(sample_id = c("QC1", "QC2"),
                 sample_name = c("QC Pool", "QC Pool"), injection_idx = c(1, 3))
@@ -17,9 +19,8 @@ sampleData(msexp) <- sd
 
 ## define file names containing spectra data for the samples and
 ## add them, along with other arbitrary files to the experiment
-fls <- dir(system.file("sciex", package = "msdata"), full.names = TRUE)
 experimentFiles(msexp) <- MsExperimentFiles(
-    mzML_files = fls,
+    mzML_files = sciex_file,
     annotations = "internal_standards.txt")
 ## link samples to data files: first sample to first file in "mzML_files",
 ## second sample to second file in "mzML_files"
@@ -28,9 +29,8 @@ msexp <- linkSampleData(msexp, with = "experimentFiles.mzML_files",
 msexp <- linkSampleData(msexp, with = "experimentFiles.annotations",
     sampleIndex = c(1, 2), withIndex = c(1, 1))
 
-library(Spectra)
 ## import the data and add it to the msexp object
-spectra(msexp) <- Spectra(fls, backend = MsBackendMzR())
+spectra(msexp) <- Spectra(sciex_file)
 
 
 qm_spectra <- c("chromatographyDuration", "ticQuantileRtFraction",